requirements.txt
nextcord>=3.0.0
python-dotenv
